Moved stuff into constants.
