+++ app/frontend/src/components/TopMenu/subcomponents/TopMenuFrame.tsx
@@ -7,7 +7,7 @@ import { type PropsWithChildren } from "react";
 import { type FC } from "react";
 import { AppBar } from "../../layouts/NewLayout/components/AppBar";
 import { useAppStore } from "../../../stores/AppStore";
-import { LEFT_MENU_WIDTH } from "../../../constants/frameDimensions";
+import { LEFT_MENU_WIDTH } from "../../../constants";
 
 type TopMenuProps = PropsWithChildren<{
     childrenLeftOfLogo?: React.ReactNode;
