--- app/frontend/src/constants/frameDimensions.ts
@@ -1,3 +1,0 @@
-// For some reason, at 300, a horizontal scrollbar appears in the left menu
-// and at 301, it disappears.
-export const LEFT_MENU_WIDTH = 301;
